extraction of zookeeper cache code modeled after accumulo-utils module.
 Changes to be committed:
	modified:   pom.xml
	modified:   services/build-parent/pom.xml
	new file:   services/zookeeper-utils/pom.xml
	renamed:    web-services/common/src/main/java/datawave/webservice/common/cache/SharedBoolean.java -> services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedBoolean.java
	renamed:    web-services/common/src/main/java/datawave/webservice/common/cache/SharedBooleanListener.java -> services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedBooleanListener.java
	renamed:    web-services/common/src/main/java/datawave/webservice/common/cache/SharedBooleanReader.java -> services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedBooleanReader.java
	renamed:    web-services/common/src/main/java/datawave/webservice/common/cache/SharedCacheCoordinator.java -> services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedCacheCoordinator.java
	renamed:    web-services/common/src/main/java/datawave/webservice/common/cache/SharedTriState.java -> services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedTriState.java
	renamed:    web-services/common/src/main/java/datawave/webservice/common/cache/SharedTriStateListener.java -> services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedTriStateListener.java
	renamed:    web-services/common/src/main/java/datawave/webservice/common/cache/SharedTriStateReader.java -> services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedTriStateReader.java
	modified:   warehouse/query-core/src/main/java/datawave/query/util/MetadataHelperCacheManagementListener.java
	modified:   warehouse/query-core/src/main/java/datawave/query/util/MetadataHelperUpdateHdfsListener.java
	modified:   web-services/common/pom.xml
	modified:   web-services/common/src/main/java/datawave/webservice/common/cache/AccumuloTableCache.java
	modified:   web-services/common/src/main/java/datawave/webservice/common/cache/BaseTableCache.java
	modified:   web-services/common/src/main/java/datawave/webservice/common/cache/TableCache.java
	modified:   web-services/common/src/test/java/datawave/webservice/common/cache/SharedCacheCoordinatorTest.java
	modified:   web-services/common/src/test/java/datawave/webservice/common/curator/TestSharedCacheCoordinator.java
