+++ pom.xml
@@ -19,6 +19,8 @@
         <module>services/accumulo-utils</module>
         <!-- TODO: Remove this when the first version of base-rest-responses is released -->
         <module>services/base-rest-responses</module>
+        <!-- TODO: Remove this when the first version of zookeeper-utils is released -->
+        <module>services/zookeeper-utils</module>
         <module>common-test</module>
         <module>web-services</module>
         <module>warehouse</module>
@@ -86,6 +88,7 @@
         <version.microservice.authorization-api>1.3</version.microservice.authorization-api>
         <version.microservice.base-rest-responses>1.0-SNAPSHOT</version.microservice.base-rest-responses>
         <version.microservice.metrics-reporter>1.2</version.microservice.metrics-reporter>
+        <version.microservice.zookeeper-utils>1.0-SNAPSHOT</version.microservice.zookeeper-utils>
         <version.minlog>1.2</version.minlog>
         <version.mysql-connector>5.1.35</version.mysql-connector>
         <version.netty>4.0.15.Final</version.netty>
@@ -310,6 +313,11 @@
                 <groupId>gov.nsa.datawave.microservice</groupId>
                 <artifactId>metrics-reporter</artifactId>
                 <version>${version.microservice.metrics-reporter}</version>
+            </dependency>
+            <dependency>
+                <groupId>gov.nsa.datawave.microservice</groupId>
+                <artifactId>zookeeper-utils</artifactId>
+                <version>${version.microservice.zookeeper-utils}</version>
             </dependency>
             <dependency>
                 <groupId>io.astefanutti.metrics.cdi</groupId>
