+++ services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedBoolean.java
@@ -3,13 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package datawave.webservice.common.cache;
+package datawave.zookeeper.cache.util;
 
 import com.google.common.collect.Maps;
 import com.google.common.util.concurrent.MoreExecutors;
 import java.io.Closeable;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.Map;
 import java.util.concurrent.Executor;
@@ -18,7 +17,6 @@ import org.apache.curator.framework.listen.Listenable;
 import org.apache.curator.framework.recipes.shared.SharedValue;
 import org.apache.curator.framework.recipes.shared.SharedValueListener;
 import org.apache.curator.framework.recipes.shared.SharedValueReader;
-import org.apache.curator.framework.recipes.shared.VersionedValue;
 import org.apache.curator.framework.state.ConnectionState;
 import org.apache.log4j.Logger;
