+++ services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedTriState.java
@@ -1,4 +1,4 @@
-package datawave.webservice.common.cache;
+package datawave.zookeeper.cache.util;
 
 import com.google.common.collect.Maps;
 import com.google.common.util.concurrent.MoreExecutors;
@@ -13,7 +13,6 @@ import org.apache.log4j.Logger;
 import java.io.Closeable;
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.Executor;
