+++ services/zookeeper-utils/src/main/java/datawave/zookeeper/cache/util/SharedTriStateReader.java
@@ -1,4 +1,4 @@
-package datawave.webservice.common.cache;
+package datawave.zookeeper.cache.util;
 
 import org.apache.curator.framework.listen.Listenable;
